backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
bcrypt==4.2.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
cssselect==1.2.0
cssutils==2.11.1
dnspython==2.7.0
email_validator==2.2.0
emails==0.6
fastapi==0.115.4
fastapi-events==0.12.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
more-itertools==10.5.0
packaging==24.1
phonenumbers==8.13.49
pluggy==1.5.0
premailer==3.10.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
